fix: add missing entity id in FsEventMapper::getFromQueue() (#196)
The FileSystemListener background job tries to delete events from the queue if the home-dir of the user cannot be found (e.g. because the user account has been removed):

https://github.com/rotdrop/nextcloud-app-context-chat/blob/f8566e58d3a384d67615163ef0ac4cf13f276746/lib/BackgroundJobs/FileSystemListenerJob.php#L104

However, this call to `FsEventMapper::delete()` cannot succeed if the entity id is missing, as is currently the case because ATM `FsEventMapper::getFromQueue()` omits the id in its result "entities" (which are in fact then only partial entities).

Another way to fix this would be to use `FsEventMapper::deleteByContent()` instead.

Signed-off-by: Claus-Justus Heine <himself@claus-justus-heine.de>